Rename formats/cp2k.py to formats/cp2klog.py
The module name was changed to better reflect the fact that it parses
CP2K output file, similar to gaussianlog.py.
