+++ iodata/test/test_cp2klog.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program; if not, see <http://www.gnu.org/licenses/>
 # --
-"""Test iodata.formats.cp2k module."""
+"""Test iodata.formats.cp2klog module."""
 
 import pytest
